requirements.txt
Django==3.0.6
Pillow==7.1.2
django-environ==0.4.5
django-ckeditor==5.9.0
django-simple-captcha==0.5.12
ipython==7.14.0
django-debug-toolbar==2.2
django-mptt==0.11.0